+++ datumbox-framework-common/src/main/java/com/datumbox/framework/common/persistentstorage/abstracts/AbstractFileDBConfiguration.java
@@ -33,7 +33,7 @@ public abstract class AbstractFileDBConfiguration implements DatabaseConfigurati
 
     /** {@inheritDoc} */
     @Override
-    public String getDBnameSeparator() {
+    public String getDBNameSeparator() {
         return File.separator;
     }